Bring details.ts and App.tsx under the size gate, prune stale exceptions
Splits presentation/details.ts by operation family (portDetails.ts for
port banner/HTTP/TLS probes, rowDetails.ts for per-kind row lines) and
extracts App.tsx's workspace region into WorkspaceView.tsx, taking the
warning-dismissal state with it. Both files' documented transition caps
were being exceeded once all v0.3.0 branches merged together.

check-file-size.mjs exceptions for files the release branches already
shrank or deleted are removed, and the useWorkspace/App caps tightened
to their new sizes so the gate ratchets instead of drifting.
