+++ apps/netscli-gui/src/tools/presentation.ts
@@ -2,17 +2,14 @@ export { columnsFor } from './presentation/columns';
 export { buildCommand } from './presentation/commands';
 export { copyContextForCell } from './presentation/copy';
 export {
-  detailLinesForRow,
   detailTabsFor,
   inspectOverviewLines,
   inspectPortsLines,
-  portBannerLines,
-  portHeaderLines,
-  portRawPreview,
-  portTlsLines,
   selectedRowsRawPreview,
   selectionSummaryLines,
 } from './presentation/details';
+export { portBannerLines, portHeaderLines, portRawPreview, portTlsLines } from './presentation/portDetails';
+export { detailLinesForRow } from './presentation/rowDetails';
 export { filterHintsFor, type FilterHints, type FilterSectionConfig } from './presentation/filterHints';
 export { latencyOf, statusOf } from './presentation/ports';
 export { buildRows } from './presentation/rows';
