+++ apps/netscli-gui/src/tools/presentation/detailLine.ts
@@ -1,0 +1,5 @@
+export interface DetailLine {
+  label: string;
+  value: string;
+  muted?: boolean;
+}
