+++ apps/netscli-gui/src/tools/presentation/details.ts
@@ -1,11 +1,9 @@
-import type { InspectResult, PortResult, PortStatus } from '../../types/netscli';
+import type { InspectResult } from '../../types/netscli';
 import type { DetailTab, ResultColumn, ResultRow } from '../types';
+import type { DetailLine } from './detailLine';
+import { httpReason, latencyOfPort, portStatusMeaning, tlsReason } from './portDetails';
 
-export interface DetailLine {
-  label: string;
-  value: string;
-  muted?: boolean;
-}
+export type { DetailLine } from './detailLine';
 
 export function detailTabsFor(row: ResultRow | undefined, selectedCount = row ? 1 : 0): DetailTab[] {
   if (selectedCount > 1) return ['selection', 'raw'];
@@ -78,58 +76,6 @@ export function selectedRowsRawPreview(rows: ResultRow[]): string {
   return JSON.stringify(rows.map((row) => row.raw), null, 2);
 }
 
-export function portBannerLines(port: PortResult, latency: string): DetailLine[] {
-  return [
-    {
-      label: 'Banner',
-      value: port.banner?.trim() || 'No plaintext banner captured',
-      muted: !port.banner?.trim(),
-    },
-    {
-      label: 'Reason',
-      value: port.banner?.trim() ? 'Application data was captured during the banner probe.' : bannerReason(port),
-      muted: !port.banner?.trim(),
-    },
-    { label: 'Service', value: port.service || inferredProtocolLabel(port) },
-    { label: 'Latency', value: latency || statusLatencyLabel(port.status) },
-    { label: 'Error', value: port.error || '-' },
-  ];
-}
-
-export function portHeaderLines(port: PortResult): DetailLine[] {
-  const headers = port.http?.headers ?? [];
-  if (port.http?.status_line || headers.length > 0) {
-    return [
-      { label: 'Status', value: port.http?.status_line || '-' },
-      ...headers.map((header) => ({ label: header.name, value: header.value })),
-    ];
-  }
-
-  return [
-    { label: 'Status', value: 'No HTTP response captured', muted: true },
-    { label: 'Reason', value: httpReason(port), muted: true },
-  ];
-}
-
-export function portTlsLines(port: PortResult): DetailLine[] {
-  if (port.tls?.protocol || port.tls?.cipher_suite || port.tls?.alpn) {
-    return [
-      { label: 'Protocol', value: port.tls?.protocol || '-' },
-      { label: 'Cipher', value: port.tls?.cipher_suite || '-' },
-      { label: 'ALPN', value: port.tls?.alpn || 'Not negotiated', muted: !port.tls?.alpn },
-    ];
-  }
-
-  return [
-    { label: 'Protocol', value: 'No TLS handshake captured', muted: true },
-    { label: 'Reason', value: tlsReason(port), muted: true },
-  ];
-}
-
-export function portRawPreview(port: PortResult): string {
-  return JSON.stringify(port, null, 2);
-}
-
 export function inspectOverviewLines(result: InspectResult): DetailLine[] {
   const ports = result.ports ?? result.open_ports;
   const open = result.open_ports.length;
@@ -176,241 +122,6 @@ export function inspectPortsLines(result: InspectResult, row: ResultRow | undefi
     { label: 'Statuses', value: formatCounts(statuses) },
     { label: 'Open Ports', value: openPorts.length > 0 ? formatList(openPorts, 16) : 'No open ports found' },
   ];
-}
-
-export function detailLinesForRow(row: ResultRow): DetailLine[] {
-  if (row.kind === 'pcap') return pcapDetailLines(row.raw);
-
-  const lines = Object.entries(row.data).map(([key, value]) => ({
-    label: key,
-    value: value == null || value === '' ? '-' : String(value),
-  }));
-
-  switch (row.kind) {
-    case 'ping':
-      return pingDetailLines(row.raw);
-    case 'trace':
-      return traceDetailLines(row.raw);
-    case 'reverse':
-      return [
-        ...lines,
-        {
-          label: 'Summary',
-          value: row.data.hostname ? 'Reverse DNS name found for this IP.' : 'No reverse DNS name returned.',
-          muted: !row.data.hostname,
-        },
-      ];
-    case 'discover':
-      return [
-        ...lines,
-        {
-          label: 'Summary',
-          value: `Responded host${row.data.vendor ? ', known vendor' : ', vendor unknown'}`,
-        },
-      ];
-    case 'sweep':
-      return [
-        ...lines,
-        {
-          label: 'Summary',
-          value: Number(row.data.open_ports || 0) > 0
-            ? `${row.data.open_ports} open service group(s) found`
-            : 'No exposed services found on scanned ports',
-        },
-      ];
-    case 'arp':
-      return [
-        { label: 'Source', value: 'Local neighbor cache' },
-        ...lines,
-      ];
-    case 'interfaces':
-      return [
-        ...lines,
-        {
-          label: 'Summary',
-          value: `${row.data.state || 'unknown'} ${row.data.kind || 'interface'}`,
-        },
-      ];
-    case 'dns':
-      return lines.filter((line) => line.value !== '-');
-    case 'mdns':
-      return mdnsDetailLines(row.raw);
-    default:
-      return lines;
-  }
-}
-
-function pingDetailLines(raw: unknown): DetailLine[] {
-  const summary = raw as Record<string, unknown>;
-  const sent = Number(summary.sent ?? 0);
-  const received = Number(summary.received ?? 0);
-  const loss = Number(summary.loss_pct ?? 0);
-  return [
-    { label: 'Host', value: valueOrDash(summary.host) },
-    { label: 'Resolved IP', value: valueOrDash(summary.ip) },
-    { label: 'Sent', value: valueOrDash(summary.sent) },
-    { label: 'Received', value: valueOrDash(summary.received) },
-    { label: 'Packet Loss', value: `${formatNumeric(loss)}%`, muted: loss > 0 },
-    { label: 'RTT Average', value: optionalMs(summary.rtt_ms_avg) },
-    { label: 'RTT Min', value: optionalMs(summary.rtt_ms_min) },
-    { label: 'RTT Max', value: optionalMs(summary.rtt_ms_max) },
-    {
-      label: 'Summary',
-      value:
-        received > 0
-          ? `${received} of ${sent} probes replied.`
-          : 'No replies were received before timeout.',
-      muted: received === 0,
-    },
-  ];
-}
-
-function traceDetailLines(raw: unknown): DetailLine[] {
-  const hop = raw as Record<string, unknown>;
-  return [
-    { label: 'Hop', value: valueOrDash(hop.hop) },
-    { label: 'Status', value: valueOrDash(hop.status) },
-    { label: 'Best RTT', value: valueOrDash(hop.best) },
-    { label: 'Average RTT', value: valueOrDash(hop.avg) },
-    { label: 'Worst RTT', value: valueOrDash(hop.worst) },
-    { label: 'Address / Host', value: valueOrDash(hop.address) },
-    { label: 'Tool', value: valueOrDash(hop.tool) },
-    { label: 'Exit Code', value: valueOrDash(hop.exit_code) },
-    { label: 'Output', value: valueOrDash(hop.line) },
-  ];
-}
-
-function mdnsDetailLines(raw: unknown): DetailLine[] {
-  const service = raw as Record<string, unknown>;
-  const properties = service.properties && typeof service.properties === 'object'
-    ? Object.entries(service.properties as Record<string, unknown>)
-        .map(([key, value]) => `${key}=${String(value)}`)
-        .join(', ')
-    : '';
-  return [
-    { label: 'Service Type', value: valueOrDash(service.service_type) },
-    { label: 'Hostname', value: valueOrDash(service.hostname) },
-    { label: 'Addresses', value: Array.isArray(service.addresses) ? service.addresses.join(', ') : '-' },
-    { label: 'Port', value: valueOrDash(service.port) },
-    { label: 'Full Name', value: valueOrDash(service.full_name) },
-    { label: 'TXT Properties', value: properties || '-', muted: !properties },
-  ];
-}
-
-function pcapDetailLines(raw: unknown): DetailLine[] {
-  const packet = raw as Record<string, unknown>;
-  const lines: DetailLine[] = [
-    { label: 'No.', value: valueOrDash(packet.index) },
-    { label: 'Timestamp', value: valueOrDash(packet.timestamp) },
-    { label: 'Protocol', value: valueOrDash(packet.protocol) },
-    { label: 'Source', value: valueOrDash(packet.source) },
-    { label: 'Destination', value: valueOrDash(packet.destination) },
-    { label: 'Length', value: valueOrDash(packet.length) },
-    { label: 'Captured', value: valueOrDash(packet.captured_length) },
-    { label: 'Info', value: valueOrDash(packet.info) },
-  ];
-
-  addOptionalLine(lines, 'Ethernet source', packet.ethernet_source);
-  addOptionalLine(lines, 'Ethernet destination', packet.ethernet_destination);
-  addOptionalLine(lines, 'Source port', packet.source_port);
-  addOptionalLine(lines, 'Destination port', packet.destination_port);
-  addOptionalLine(lines, 'TCP flags', packet.tcp_flags);
-  addOptionalLine(lines, 'ICMP type', packet.icmp_type);
-  addOptionalLine(lines, 'ICMP code', packet.icmp_code);
-  addOptionalLine(lines, 'ARP operation', packet.arp_operation);
-  addOptionalLine(lines, 'Hex preview', packet.hex_preview);
-
-  return lines;
-}
-
-function addOptionalLine(lines: DetailLine[], label: string, value: unknown) {
-  if (value == null || value === '') return;
-  lines.push({ label, value: String(value) });
-}
-
-function valueOrDash(value: unknown): string {
-  return value == null || value === '' ? '-' : String(value);
-}
-
-function optionalMs(value: unknown): string {
-  return typeof value === 'number' && Number.isFinite(value) ? `${formatNumeric(value)} ms` : '-';
-}
-
-function formatNumeric(value: number): string {
-  return Number.isInteger(value) ? String(value) : value.toFixed(1);
-}
-
-function bannerReason(port: PortResult): string {
-  switch (port.status) {
-    case 'open':
-      return 'The TCP connection opened, but the service did not send plaintext before the probe timeout.';
-    case 'closed':
-      return 'The connection was refused, so there was no application session to probe.';
-    case 'filtered':
-      return 'The TCP connect attempt timed out before application data could be read.';
-    case 'error':
-      return port.error || 'The probe failed before banner capture completed.';
-  }
-}
-
-function httpReason(port: PortResult): string {
-  if (port.status !== 'open') {
-    return 'HTTP probing is skipped unless the TCP connection opens first.';
-  }
-  if (!isHttpLike(port)) {
-    return 'This port/service is not HTTP-like, so an HTTP HEAD probe was not attempted.';
-  }
-  return 'The HTTP probe did not receive a status line or headers before its timeout.';
-}
-
-function tlsReason(port: PortResult): string {
-  if (port.status !== 'open') {
-    return 'TLS probing is skipped unless the TCP connection opens first.';
-  }
-  if (!isTlsLike(port)) {
-    return 'This port/service is not TLS-like, so a TLS handshake probe was not attempted.';
-  }
-  return 'The TLS handshake did not complete before the probe timeout.';
-}
-
-function isHttpLike(port: PortResult): boolean {
-  const service = port.service?.toLowerCase() ?? '';
-  return service.includes('http') || [80, 443, 8000, 8008, 8080, 8443, 8888].includes(port.port);
-}
-
-function isTlsLike(port: PortResult): boolean {
-  const service = port.service?.toLowerCase() ?? '';
-  return service.includes('https') || service.includes('tls') || service.includes('ssl') || [443, 8443].includes(port.port);
-}
-
-function inferredProtocolLabel(port: PortResult): string {
-  if (isHttpLike(port)) return 'http-like';
-  if (isTlsLike(port)) return 'tls-like';
-  return 'tcp';
-}
-
-function statusLatencyLabel(status: PortStatus): string {
-  if (status === 'filtered') return 'timeout';
-  if (status === 'closed') return 'refused';
-  return '-';
-}
-
-function latencyOfPort(port: PortResult): string {
-  if (typeof port.latency_ms === 'number') return `${port.latency_ms} ms`;
-  return statusLatencyLabel(port.status);
-}
-
-function portStatusMeaning(port: PortResult): string {
-  switch (port.status) {
-    case 'open':
-      return 'TCP connect succeeded.';
-    case 'closed':
-      return 'The host refused the connection.';
-    case 'filtered':
-      return 'The probe timed out or was blocked before connect.';
-    case 'error':
-      return port.error || 'The probe failed before classification completed.';
-  }
 }
 
 function countValues(values: string[]): Record<string, number> {
